+++ src/main/kotlin/repositories/JugadorRepository.kt
@@ -1,6 +1,5 @@
 package org.example.repositories
 
 import org.example.models.Jugador
-import org.example.repository.CrudRepository
 
-interface JugadorRepository : CrudRepository<Jugador, Int>
+interface JugadorRepository : CrudRepository<Jugador, Long>
